Cast pointers as long to silence compiler warning  On most 64-bit architectures, casting pointer expression to int means cutting  off high-order bits. There are special types intptr_t and ptrdiff_t for it. Author: Sergey Romanov <schragge@gmail.com>

+++ session.h
@@ -91,10 +91,10 @@ typedef union
 	char	sizetester[BLKSIZE];	/* forces BLK to be prefered size */
 } BLK;
 
-#define SES_MAXSUPER	((o_blksize - (int)(((BLK *)0)->super.buf)) / sizeof(BLKNO))
-#define SES_MAXSUPER2	((o_blksize - (int)(((BLK *)0)->super2.buf)) / sizeof(BLKNO))
-#define SES_MAXBUFINFO	(o_blksize - (int)(((BLK *)0)->bufinfo.name))
-#define SES_MAXBLKLIST	((o_blksize - (int)(((BLK *)0)->blklist.blk)) / sizeof(((BLK *)0)->blklist.blk[0]))
+#define SES_MAXSUPER	((o_blksize - (long)(((BLK *)0)->super.buf)) / sizeof(BLKNO))
+#define SES_MAXSUPER2	((o_blksize - (long)(((BLK *)0)->super2.buf)) / sizeof(BLKNO))
+#define SES_MAXBUFINFO	(o_blksize - (long)(((BLK *)0)->bufinfo.name))
+#define SES_MAXBLKLIST	((o_blksize - (long)(((BLK *)0)->blklist.blk)) / sizeof(((BLK *)0)->blklist.blk[0]))
 #define SES_MAXCHARS	(o_blksize / sizeof(CHAR))
